Code for connecting to peers
