+++ app/src/main/java/com/finastra/payments/payoff/MainActivity.java
@@ -1,20 +1,7 @@
 package com.finastra.payments.payoff;
 
-import android.content.BroadcastReceiver;
-import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
-import android.net.wifi.WifiManager;
-import android.net.wifi.p2p.WifiP2pDevice;
-import android.net.wifi.p2p.WifiP2pDeviceList;
-import android.net.wifi.p2p.WifiP2pManager.*;
-import android.net.wifi.p2p.WifiP2pManager;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
@@ -25,40 +12,18 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.ArrayAdapter;
-import android.widget.LinearLayout;
-import android.widget.ListView;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.finastra.payments.wifidirect.WiFiDirectBroadcastReceiver;
-
-import java.util.ArrayList;
-import java.util.List;
+import android.widget.Spinner;
 
 
 public class MainActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener, WifiP2pManager.PeerListListener {
+        implements NavigationView.OnNavigationItemSelectedListener{
 
-    //Instance variables
-    private LinearLayout parentLinearLayout;
-
-    WifiP2pManager wifiP2pManager;
-    Channel channel;
-    BroadcastReceiver broadcastReceiver;
-    IntentFilter intentFilter;
-    ListView discoveredPeersList;
-    WifiManager wifiManager;
-
-    List<WifiP2pDevice> peers = new ArrayList<WifiP2pDevice>();
-    String[] deviceNameArray;
-    WifiP2pDevice[] deviceArray;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         init();
-        initWifiDirectComponents();
     }
 
     public void init() {
@@ -66,14 +31,14 @@ public class MainActivity extends AppCompatActivity
         startActivity(intent);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-            }
-        });
+
+        Spinner spinner = findViewById(R.id.spr_currencies);
+        ArrayAdapter<CharSequence> adapter = ArrayAdapter.createFromResource(this,
+                R.array.currency_array, android.R.layout.simple_spinner_item);
+        // Specify the layout to use when the list of choices appears
+        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+        // Apply the adapter to the spinner
+        spinner.setAdapter(adapter);
 
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
         ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
@@ -83,21 +48,8 @@ public class MainActivity extends AppCompatActivity
 
         NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
-        parentLinearLayout = findViewById(R.id.discoveredPeers);
     }
 
-    public void initWifiDirectComponents() {
-        wifiManager = (WifiManager) getApplicationContext().getSystemService(Context.WIFI_SERVICE);
-        wifiP2pManager = (WifiP2pManager) getSystemService(Context.WIFI_P2P_SERVICE);
-        channel = wifiP2pManager.initialize(this,getMainLooper(),null);
-        broadcastReceiver = new WiFiDirectBroadcastReceiver(wifiP2pManager,channel,this);
-        wifiManager.setWifiEnabled(true);
-        intentFilter = new IntentFilter();
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_STATE_CHANGED_ACTION);
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_PEERS_CHANGED_ACTION);
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_CONNECTION_CHANGED_ACTION);
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_THIS_DEVICE_CHANGED_ACTION);
-    }
 
     @Override
     public void onBackPressed() {
@@ -149,7 +101,6 @@ public class MainActivity extends AppCompatActivity
 
         } else if (id == R.id.navLogut) {
             Intent intent = new Intent(MainActivity.this, LoginActivity.class);
-            wifiManager.setWifiEnabled(false);
             startActivity(intent);
         }
 
@@ -159,95 +110,11 @@ public class MainActivity extends AppCompatActivity
     }
 
     public void onDelete(View v) {
-        parentLinearLayout.removeView((View) v.getParent());
+        //parentLinearLayout.removeView((View) v.getParent());
     }
 
-    public void onSendMoney(View v) {
-        TextView amountToSend = findViewById(R.id.idInputAmount);
-        TextView eWalletAmount = findViewById(R.id.eWalletAmount);
-        String amountToSendStr = amountToSend.getText() + "";
-        Log.i("AMOUNT TO SEND: ",amountToSendStr);
-        if(!amountToSendStr.equals("")) {
-            String eWalletAmountStr = eWalletAmount.getText() +"";
-            double oldValue = Double.parseDouble(eWalletAmountStr);
-            double subtractedValue = Double.parseDouble(amountToSendStr);
-            if (subtractedValue > oldValue) {
-                Context context = getApplicationContext();
-                CharSequence text = "Amount to send is greater than the balance";
-                int duration = Toast.LENGTH_SHORT;
-                Toast.makeText(context, text, duration).show();
-            } else {
-                double newBalance = oldValue - subtractedValue;
-                eWalletAmount.setText(newBalance + "");
-            }
-        } else {
-            Context context = getApplicationContext();
-            CharSequence text = "Invalid amount";
-            int duration = Toast.LENGTH_SHORT;
-            Toast.makeText(context, text, duration).show();
-        }
-    }
-
-    //WiFiDirect
-
-    public void onDiscover(View v) {
-        if(wifiManager.isWifiEnabled()) {
-            wifiP2pManager.discoverPeers(channel,new WifiP2pManager.ActionListener(){
-                @Override
-                public void onSuccess() {
-                    Log.i("Main:","Discovery started");
-                }
-
-                @Override
-                public void onFailure(int i) {
-                    Toast.makeText(getApplicationContext(), "Discovery starting failed", Toast.LENGTH_SHORT).show();
-                    //connectionStatus.setText("Discovery starting failed");
-                }
-            });
-        } else {
-            wifiManager.setWifiEnabled(true);
-        }
-    }
-
-
-    @Override
-    public void onPeersAvailable(WifiP2pDeviceList peerlist) {
-        if (!peerlist.getDeviceList().equals(peers)) {
-            peers.clear();
-            peers.addAll(peerlist.getDeviceList());
-            deviceNameArray = new String[peerlist.getDeviceList().size()];
-            deviceArray = new WifiP2pDevice[peerlist.getDeviceList().size()];
-            Log.i("DEVICES NAME: ", String.valueOf(peerlist.getDeviceList().size()));
-            int index = 0;
-            for (WifiP2pDevice device : peerlist.getDeviceList()) {
-                deviceNameArray[index] = device.deviceName;
-
-                deviceArray[index] = device;
-                index++;
-                LayoutInflater inflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-                View rowView = inflater.inflate(R.layout.field, null);
-                TextView textView = rowView.findViewById(R.id.number_edit_text);
-                textView.setText(device.deviceName);
-                // Add the new row before the add field button.
-                parentLinearLayout.addView(rowView, parentLinearLayout.getChildCount() - 1);
-            }
-        }
-        if (peers.size() == 0) {
-            Toast.makeText(getApplicationContext(), "No device found", Toast.LENGTH_SHORT).show();
-            return;
-        }
-    }
-
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        registerReceiver(broadcastReceiver,intentFilter);
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        unregisterReceiver(broadcastReceiver);
+    public void onSend (View v) {
+        Intent intent = new Intent(MainActivity.this, SendPaymentActivity.class);
+        startActivity(intent);
     }
 }
