+++ app/src/main/java/com/finastra/payments/wifidirect/WiFiDirectBroadcastReceiver.java
@@ -11,26 +11,25 @@ import android.net.wifi.p2p.WifiP2pManager.*;
 import android.os.Build;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
-import android.util.Log;
 import android.widget.Toast;
 
-import com.finastra.payments.payoff.MainActivity;
+import com.finastra.payments.payoff.SendPaymentActivity;
 
 public class WiFiDirectBroadcastReceiver extends BroadcastReceiver {
 
     private WifiP2pManager wifiP2pManager;
     private Channel channel;
-    private MainActivity mainActivity;
+    private SendPaymentActivity sendPaymentActivity;
     PeerListListener myPeerListListener;
     private static final int PERMISSIONS_REQUEST_CODE_ACCESS_COARSE_LOCATION = 1001;
 
 
     public WiFiDirectBroadcastReceiver(WifiP2pManager wifiP2pManager, WifiP2pManager.Channel channel,
-                                       MainActivity activity) {
+                                       SendPaymentActivity activity) {
         super();
         this.wifiP2pManager = wifiP2pManager;
         this.channel = channel;
-        this.mainActivity = activity;
+        this.sendPaymentActivity = activity;
     }
 
     @Override
@@ -48,11 +47,11 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver {
             // Call WifiP2pManager.requestPeers() to get a list of current peers
             Toast.makeText(context,"Requesting peers",Toast.LENGTH_SHORT).show();
             if (wifiP2pManager != null) {
-                if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && ContextCompat.checkSelfPermission(mainActivity,Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED){
-                   ActivityCompat.requestPermissions(mainActivity,new String[]{Manifest.permission.ACCESS_COARSE_LOCATION}, PERMISSIONS_REQUEST_CODE_ACCESS_COARSE_LOCATION);
+                if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && ContextCompat.checkSelfPermission(sendPaymentActivity,Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED){
+                   ActivityCompat.requestPermissions(sendPaymentActivity,new String[]{Manifest.permission.ACCESS_COARSE_LOCATION}, PERMISSIONS_REQUEST_CODE_ACCESS_COARSE_LOCATION);
                     //After this point you wait for callback in onRequestPermissionsResult(int, String[], int[]) overriden method
                 } else{
-                    wifiP2pManager.requestPeers(channel, mainActivity);
+                    wifiP2pManager.requestPeers(channel, sendPaymentActivity);
                     //do something, permission was previously granted; or legacy device
                 }
             }
@@ -65,7 +64,7 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver {
             if (networkInfo.isConnected()) {
                 // we are connected with the other device, request connection
                 // info to find group owner IP
-                wifiP2pManager.requestConnectionInfo(channel, mainActivity);
+                wifiP2pManager.requestConnectionInfo(channel, sendPaymentActivity);
             } else if (WifiP2pManager.WIFI_P2P_THIS_DEVICE_CHANGED_ACTION.equals(action)) {
                 // Respond to this device's wifi state changing
                 Toast.makeText(context,"Disconnected",Toast.LENGTH_SHORT).show();
