--- app/src/main/res/layout-land/content_main.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/idMainLayout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.finastra.payments.payoff.MainActivity"
-    tools:showIn="@layout/app_bar_main">
-
-    <Button
-        android:id="@+id/id_btnOnDiscover"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/buttonDiscover"
-        app:layout_constraintBottom_toTopOf="@+id/scrollDiscoveredPeers"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.707"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/idInputAmount" />
-
-    <Button
-        android:id="@+id/id_btnRequest"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/buttonRequest"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/id_btnSend"
-        app:layout_constraintTop_toBottomOf="@+id/scrollDiscoveredPeers" />
-
-    <Button
-        android:id="@+id/id_btnSend"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/buttonSend"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/id_btnRequest"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/scrollDiscoveredPeers" />
-
-    <EditText
-        android:id="@+id/idInputAmount"
-        android:layout_width="197dp"
-        android:layout_height="48dp"
-        android:importantForAutofill="no"
-        android:inputType="phone"
-        android:text="@string/inputAmount"
-        android:textColor="@color/colorPrimary"
-        android:textColorHint="@color/colorPrimary"
-        android:textStyle="normal|bold|italic"
-        app:layout_constraintBottom_toTopOf="@+id/id_btnOnDiscover"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/wifiOnOff"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:targetApi="o" />
-
-    <Switch
-        android:id="@+id/wifiOnOff"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/wifiOnOff"
-        app:layout_constraintBottom_toTopOf="@+id/scrollDiscoveredPeers"
-        app:layout_constraintEnd_toStartOf="@+id/idInputAmount"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.247" />
-
-    <ScrollView
-        android:id="@+id/scrollDiscoveredPeers"
-        android:layout_width="353dp"
-        android:layout_height="250dp"
-        app:layout_constraintBottom_toTopOf="@+id/id_btnRequest"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.509"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/id_btnOnDiscover">
-
-        <LinearLayout
-            android:id="@+id/discoveredPeers"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical" />
-    </ScrollView>
-
-</android.support.constraint.ConstraintLayout>
