+++ app/src/main/res/layout-v26/content_main.xml
@@ -11,111 +11,64 @@
     tools:showIn="@layout/app_bar_main">
 
     <Button
-        android:id="@+id/id_btnOnDiscover"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:onClick="onDiscover"
-        android:text="@string/buttonDiscover"
-        app:layout_constraintBottom_toTopOf="@+id/scrollDiscoveredPeers"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.714"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/idInputAmount" />
-
-    <Button
-        android:id="@+id/id_btnRequest"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:id="@+id/id_btnOnRequest"
+        android:layout_width="150dp"
+        android:layout_height="65dp"
+        android:onClick="onSend"
         android:text="@string/buttonRequest"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/id_btnSend"
-        app:layout_constraintTop_toBottomOf="@+id/scrollDiscoveredPeers" />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/id_btnSend" />
 
     <Button
         android:id="@+id/id_btnSend"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:onClick="onSendMoney"
+        android:layout_width="153dp"
+        android:layout_height="61dp"
+        android:onClick="onSend"
         android:text="@string/buttonSend"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/id_btnRequest"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/scrollDiscoveredPeers" />
-
-    <EditText
-        android:id="@+id/idInputAmount"
-        android:layout_width="146dp"
-        android:layout_height="43dp"
-        android:importantForAutofill="no"
-        android:inputType="phone"
-        android:text="@string/defaultValueForInputAmount"
-        android:textColor="@color/colorPrimary"
-        android:textColorHint="@color/colorPrimary"
-        android:textStyle="normal|bold|italic"
-        app:layout_constraintBottom_toTopOf="@+id/id_btnOnDiscover"
+        app:layout_constraintBottom_toTopOf="@+id/id_btnOnRequest"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/textView2"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:targetApi="o" />
-
-    <ScrollView
-        android:id="@+id/scrollDiscoveredPeers"
-        android:layout_width="353dp"
-        android:layout_height="250dp"
-        app:layout_constraintBottom_toTopOf="@+id/id_btnRequest"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.509"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/id_btnOnDiscover">
-
-        <LinearLayout
-            android:id="@+id/discoveredPeers"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical" />
-    </ScrollView>
+        app:layout_constraintTop_toBottomOf="@+id/availableBal" />
 
     <TextView
-        android:id="@+id/eWallet"
-        android:layout_width="wrap_content"
-        android:layout_height="22dp"
-        android:text="@string/eWallet"
+        android:id="@+id/idAvailableBalance"
+        android:layout_width="243dp"
+        android:layout_height="42dp"
+        android:gravity="center"
+        android:text="@string/lbl_availableBalance"
+        android:textSize="20sp"
         android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/eWalletAmount"
+        app:layout_constraintBottom_toTopOf="@+id/spr_currencies"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.188"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView2" />
-
-    <TextView
-        android:id="@+id/eWalletAmount"
-        android:layout_width="112dp"
-        android:layout_height="34dp"
-        android:text="@string/eWalletAmount"
-        android:textAlignment="center"
-        android:textColor="@color/colorAccent"
-        android:textSize="30sp"
-        app:layout_constraintBottom_toTopOf="@+id/scrollDiscoveredPeers"
-        app:layout_constraintEnd_toStartOf="@+id/id_btnOnDiscover"
-        app:layout_constraintHorizontal_bias="0.323"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/eWallet"
-        app:layout_constraintVertical_bias="0.326" />
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/lbl_amtToSend"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/eWallet"
-        app:layout_constraintEnd_toStartOf="@+id/idInputAmount"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/availableBal"
+        android:layout_width="336dp"
+        android:layout_height="148dp"
+        android:text="@string/availableBal"
+        android:textAlignment="center"
+        android:textColor="@color/colorAccent"
+        android:textSize="125sp"
+        app:layout_constraintBottom_toTopOf="@+id/id_btnSend"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/spr_currencies" />
+
+    <Spinner
+        android:id="@+id/spr_currencies"
+        android:layout_width="243dp"
+        android:layout_height="41dp"
+        app:layout_constraintBottom_toTopOf="@+id/availableBal"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/idAvailableBalance" />
 
 </android.support.constraint.ConstraintLayout>
