+++ app/src/main/res/layout/activity_send.xml
@@ -1,0 +1,88 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.finastra.payments.payoff.SendPaymentActivity">
+
+    <TextView
+        android:id="@+id/lbl_balance"
+        android:layout_width="89dp"
+        android:layout_height="44dp"
+        android:gravity="center"
+        android:text="@string/lbl_availableBalance"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toTopOf="@+id/scrollDiscoveredPeers"
+        app:layout_constraintEnd_toStartOf="@+id/idAvailableBalance"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/lbl_amount" />
+
+    <ScrollView
+        android:id="@+id/scrollDiscoveredPeers"
+        android:layout_width="353dp"
+        android:layout_height="315dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.6"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/idAvailableBalance">
+
+        <LinearLayout
+            android:id="@+id/discoveredPeers"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical" />
+    </ScrollView>
+
+    <EditText
+        android:id="@+id/idAmountToSend"
+        android:layout_width="146dp"
+        android:layout_height="57dp"
+        android:gravity="center"
+        android:importantForAutofill="no"
+        android:inputType="phone"
+        android:text="@string/defaultValueForInputAmount"
+        android:textColor="@color/colorPrimary"
+        android:textColorHint="@color/colorPrimary"
+        android:textSize="30sp"
+        android:textStyle="normal|bold|italic"
+        app:layout_constraintBottom_toTopOf="@+id/idAvailableBalance"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toEndOf="@+id/lbl_amount"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:targetApi="o" />
+
+    <TextView
+        android:id="@+id/idAvailableBalance"
+        android:layout_width="141dp"
+        android:layout_height="48dp"
+        android:gravity="center"
+        android:text="@string/availableBal"
+        android:textColor="@color/colorPrimary"
+        android:textColorHint="@color/colorPrimary"
+        android:textSize="30sp"
+        android:textStyle="normal|bold|italic"
+        app:layout_constraintBottom_toTopOf="@+id/scrollDiscoveredPeers"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toEndOf="@+id/lbl_balance"
+        app:layout_constraintTop_toBottomOf="@+id/idAmountToSend" />
+
+    <TextView
+        android:id="@+id/lbl_amount"
+        android:layout_width="89dp"
+        android:layout_height="44dp"
+        android:gravity="center"
+        android:text="@string/lbl_amtToSend"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toTopOf="@+id/lbl_balance"
+        app:layout_constraintEnd_toStartOf="@+id/idAmountToSend"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+
+</android.support.constraint.ConstraintLayout>
