+++ app/src/main/res/layout/field.xml
@@ -1,18 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-    <EditText
-        android:id="@+id/number_edit_text"
+
+    <TextView
+        android:id="@+id/discoveredPeer"
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_weight="5"
-        android:inputType=""/>
+        android:clickable="true"
+        android:onClick="onConnectToPeer" />
+
     <Button
-        android:id="@+id/id_btnDelete"
-        android:text="@string/delete"
+        android:id="@+id/id_peerSend"
         android:layout_width="0dp"
         android:layout_height="40dp"
         android:layout_weight="1"
-        android:onClick="onDelete"/>
+        android:onClick="onSendPayment"
+        android:text="@string/buttonSend" />
 </LinearLayout>
