+++ app/src/main/res/values-v21/strings.xml
@@ -8,13 +8,12 @@
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
     <string name="buttonOnWifi">WiFiOn</string>
-    <string name="buttonDiscover">Discover</string>
     <string name="buttonSend">SEND</string>
     <string name="inputAmount">Enter Amount Here</string>
     <string name="buttonRequest">REQUEST</string>
     <string name="wifiOnOff">Wifi</string>
-    <string name="eWallet">Balance:</string>
-    <string name="eWalletAmount">500</string>
+    <string name="availableBal">0.0</string>
+    <string name="lbl_availableBalance">Balance</string>
     <string name="delete">Del</string>
     <string name="lbl_amtToSend">Amount:</string>
     <string name="logout">Logout</string>
@@ -38,4 +37,15 @@
     <string name="lblFxChange">FX Exchange Rate</string>
     <string name="defaultValueForInputAmount">0.0</string>
     <string name="defaultLogin">user1@gmail.com</string>
+    <string-array name="currency_array">
+        <item>PHP</item>
+        <item>EUR</item>
+        <item>JPN</item>
+        <item>GBP</item>
+        <item>USD</item>
+        <item>AUD</item>
+        <item>CHF</item>
+        <item>CAD</item>
+        <item>HKG</item>
+    </string-array>
 </resources>
