+++ app/src/main/res/values/strings.xml
@@ -7,13 +7,12 @@
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
     <string name="buttonOnWifi">WiFiOn</string>
-    <string name="buttonDiscover">Discover</string>
     <string name="buttonSend">SEND</string>
     <string name="inputAmount">Enter Amount Here</string>
     <string name="buttonRequest">REQUEST</string>
     <string name="wifiOnOff">Wifi</string>
-    <string name="eWallet">Balance:</string>
-    <string name="eWalletAmount">500</string>
+    <string name="lbl_availableBalance">Available Balance</string>
+    <string name="availableBalance">0.0</string>
     <string name="delete">Del</string>
     <string name="lbl_amtToSend">Amount:</string>
     <string name="logout">Logout</string>
@@ -37,4 +36,15 @@
     <string name="lblFxChange">FX Exchange Rate</string>
     <string name="defaultValueForInputAmount">0.0</string>
     <string name="defaultLogin">user1@gmail.com</string>
+    <string-array name="currency_array">
+        <item>PHP</item>
+        <item>EUR</item>
+        <item>JPN</item>
+        <item>GBP</item>
+        <item>USD</item>
+        <item>AUD</item>
+        <item>CHF</item>
+        <item>CAD</item>
+        <item>HKG</item>
+    </string-array>
 </resources>
